Move inputmedia to single file + fix their thumbs
